+++ src/components/Contact.jsx
@@ -5,7 +5,7 @@ import github from '../assets/github.svg'
 function Contact() {return(
     <div className='margin'>
         <h2 id='contactSection'>Contact Me</h2>
-        <h4 >Click an icon to contact me.</h4>
+        <h4>Click an icon to contact me.</h4>
         <div id='iconStyles'>
             <div className='iconRow'>
                 <a href="https://mail.google.com/mail/?view=cm&fs=1&to=lizajparr@gmail.com" target='_blank'><img className='icon' src={gmail} alt="gmail icon" /></a>
